gs-advisor-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
langchain
langchain-text-splitters
langchain-google-genai
python-dotenv
pydantic-settings